XVIDEO: Pass image format to copy_buffer.
Since the purpose of having copy_buffer is to allow changing the
data layout when copying it, it is necessary to know what's the
input format. The xvideo layer calls copy_buffer before any other
driver function, so it is necessary to add a format parameter.

Previously the RK XV driver was using the u_offset and v_offset
parameters passed to set_yuv420_input_buffer to determine the
input format, but the first call to copy_buffer would use
potentially invalid values, ocasionanlly causing a segfault.